refactor: move WebRequest transport to its own file
